swap ease-in and ease-out
Previously, in and out were swapped with convention as it was based on
the existing approach where the keyframe interpolation type applies only
to the the time at the keyframe and later. Also, it considered "in" to
be beginning of the time segment between keyframes and "out" to be the
end. However, ease-in is supposed to affect the time before the
keyframe, and ease-out the time after the keyframe.